+++ src/main/java/com/example/yeobee/core/country/domain/Country.java
@@ -1,7 +1,6 @@
 package com.example.yeobee.core.country.domain;
 
 import com.example.yeobee.core.currency.domain.CountryCurrency;
-import com.example.yeobee.core.tripCountry.domain.TripCountry;
 import jakarta.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
