+++ src/main/java/com/example/yeobee/core/country/domain/TripCountryRepository.java
@@ -1,6 +1,7 @@
-package com.example.yeobee.core.tripCountry.domain;
+package com.example.yeobee.core.country.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TripCountryRepository extends JpaRepository<TripCountry, Long> {
+
 }
