+++ src/main/java/com/example/yeobee/core/expense/domain/UserExpenseRepository.java
@@ -1,6 +1,7 @@
-package com.example.yeobee.core.userExpense.domain;
+package com.example.yeobee.core.expense.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserExpenseRepository extends JpaRepository<UserExpense, Long> {
+
 }
