+++ src/main/java/com/example/yeobee/core/trip/domain/Trip.java
@@ -1,10 +1,9 @@
 package com.example.yeobee.core.trip.domain;
 
 import com.example.yeobee.common.entity.BaseEntity;
+import com.example.yeobee.core.country.domain.TripCountry;
 import com.example.yeobee.core.currency.domain.TripCurrency;
 import com.example.yeobee.core.expense.domain.Expense;
-import com.example.yeobee.core.tripCountry.domain.TripCountry;
-import com.example.yeobee.core.tripUser.domain.TripUser;
 import jakarta.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
