+++ src/main/java/com/example/yeobee/core/trip/domain/TripUserRepository.java
@@ -1,6 +1,7 @@
-package com.example.yeobee.core.tripUser.domain;
+package com.example.yeobee.core.trip.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TripUserRepository extends JpaRepository<TripUser, Long> {
+
 }
