+++ src/main/resources/application.yml
@@ -7,7 +7,7 @@ spring:
   jpa:
     generate-ddl: true
     hibernate:
-      ddl-auto: update
+      ddl-auto: validate
     show-sql: true
     properties:
       hibernate:
@@ -26,5 +26,5 @@ spring:
       cdn:
         url: https://cdn.dev.yeobee.me/
   flyway:
-    enabled: false
-    baselineOnMigrate: false
+    enabled: true
+    baselineOnMigrate: true
